List block fix, rewrote tests for Markdown support
