+++ src/block.js
@@ -169,12 +169,11 @@ Object.assign(Block.prototype, SimpleBlock.fn, require('./block-validations'), {
     /* Simple to start. Add conditions later */
     if (this.hasTextBlock()) {
       data.text = this.getTextBlockHTML();
+      data.isHtml = true;
 
       if (this.options.convertToMarkdown) {
         data.isHtml = false;
-        if (data.text.length) {data.text = stToMarkdown(data.text, this.type); }
-      } else {
-        data.isHtml = true;
+        if (data.text.length) {data.text = stToMarkdown(data.text, this.type);}
       }
     }
